Extract tooltip to separate file
